[ADD] Added window resize handling
[+] Added pygame's WINDOW* events

[+] Window: Added 'resizable' argument

[+] Window: Added set_size, set_width and set_height methods

[+] Scene: Added __theme_init__ method, called when a window is opening

+++ py_diamond/graphics/renderer.py
@@ -236,7 +236,7 @@ class SurfaceRenderer(Renderer):
         target: Surface = self.__target
         return Rect.convert(target.get_rect(**kwargs))
 
-    def get_size(self, /) -> Tuple[float, float]:
+    def get_size(self, /) -> Tuple[int, int]:
         target: Surface = self.__target
         return target.get_size()
